Update 3.6.c
changed oneline

+++ programming exercise/3.6.c
@@ -1,23 +1,12 @@
 #include <stdio.h>
 int main()
 {
-    int Depreciation,Purchase_price, Salvage_price, Year_of_service;
+    float Depreciation,Purchase_price, Salvage_price, Year_of_service;
 
     // main equation
     Depreciation = (Purchase_price - Salvage_price) / Year_of_service;
-    // => (Depreciation * Year_of_service) = Purchase_price - Salvage_price
-    // => Salvage_price = Purchase_price - (Depreciation * Year_of_service);
-    // => Salvage_price = -1 *((Depreciation * Year_of_service) - Purchase_price)
-    // => Salvage_price = 
-
-
-
-
-
     // modify equation
     // Salvage_price = Purchase_price - (Depreciation * Year_of_service);
-    // =>
-    
-
+    // এখন ডান পাসের ৩টা রাশির মান নিয়ে calculation করলেই মান পাওয়া যাবে
     return 0;
 }
